Reversed certbot tests if pre/post and deploy scripts are needed.
Swapped test logic, so pre was checking for deploy actions and vice
versa.  In the cloud config, both are used so I didn't notice.  The
webmail only uses one so noticed when the cert wasn't renewing.

Also refined the group vars for webmail to pull and rebuild the docker
if necessary while we have it down.

Signed-off-by: David Loffredo <loffredo@steptools.com>